ran prettier on src folder

+++ src/user/email.js
@@ -1,4 +1,3 @@
-
 'use strict';
 
 const nconf = require('nconf');
@@ -62,10 +61,7 @@ UserEmail.getValidationExpiry = async (uid) => {
 
 UserEmail.expireValidation = async (uid) => {
     const code = await db.get(`confirm:byUid:${uid}`);
-    await db.deleteAll([
-        `confirm:byUid:${uid}`,
-        `confirm:${code}`,
-    ]);
+    await db.deleteAll([`confirm:byUid:${uid}`, `confirm:${code}`]);
 };
 
 UserEmail.canSendValidation = async (uid, email) => {
@@ -90,7 +86,9 @@ UserEmail.sendValidationEmail = async function (uid, options) {
      */
 
     if (meta.config.sendValidationEmail !== 1) {
-        winston.verbose(`[user/email] Validation email for uid ${uid} not sent due to config settings`);
+        winston.verbose(
+            `[user/email] Validation email for uid ${uid} not sent due to config settings`
+        );
         return;
     }
 
@@ -116,8 +114,13 @@ UserEmail.sendValidationEmail = async function (uid, options) {
         return;
     }
 
-    if (!options.force && !await UserEmail.canSendValidation(uid, options.email)) {
-        throw new Error(`[[error:confirm-email-already-sent, ${emailConfirmInterval}]]`);
+    if (
+        !options.force &&
+        !(await UserEmail.canSendValidation(uid, options.email))
+    ) {
+        throw new Error(
+            `[[error:confirm-email-already-sent, ${emailConfirmInterval}]]`
+        );
     }
 
     const username = await user.getUserField(uid, 'username');
@@ -125,7 +128,10 @@ UserEmail.sendValidationEmail = async function (uid, options) {
         uid,
         username,
         confirm_link,
-        confirm_code: await plugins.hooks.fire('filter:user.verify.code', confirm_code),
+        confirm_code: await plugins.hooks.fire(
+            'filter:user.verify.code',
+            confirm_code
+        ),
         email: options.email,
 
         subject: options.subject || '[[email:email.verify-your-email.subject]]',
@@ -134,15 +140,23 @@ UserEmail.sendValidationEmail = async function (uid, options) {
 
     await UserEmail.expireValidation(uid);
     await db.set(`confirm:byUid:${uid}`, confirm_code);
-    await db.pexpire(`confirm:byUid:${uid}`, emailConfirmExpiry * 24 * 60 * 60 * 1000);
+    await db.pexpire(
+        `confirm:byUid:${uid}`,
+        emailConfirmExpiry * 24 * 60 * 60 * 1000
+    );
 
     await db.setObject(`confirm:${confirm_code}`, {
         email: options.email.toLowerCase(),
         uid: uid,
     });
-    await db.pexpire(`confirm:${confirm_code}`, emailConfirmExpiry * 24 * 60 * 60 * 1000);
+    await db.pexpire(
+        `confirm:${confirm_code}`,
+        emailConfirmExpiry * 24 * 60 * 60 * 1000
+    );
 
-    winston.verbose(`[user/email] Validation email for uid ${uid} sent to ${options.email}`);
+    winston.verbose(
+        `[user/email] Validation email for uid ${uid} sent to ${options.email}`
+    );
     events.log({
         type: 'email-confirmation-sent',
         uid,
@@ -166,7 +180,10 @@ UserEmail.confirmByCode = async function (code, sessionId) {
     }
 
     // If another uid has the same email, remove it
-    const oldUid = await db.sortedSetScore('email:uid', confirmObj.email.toLowerCase());
+    const oldUid = await db.sortedSetScore(
+        'email:uid',
+        confirmObj.email.toLowerCase()
+    );
     if (oldUid) {
         await UserEmail.remove(oldUid, sessionId);
     }
@@ -182,7 +199,11 @@ UserEmail.confirmByCode = async function (code, sessionId) {
     await Promise.all([
         UserEmail.confirmByUid(confirmObj.uid),
         db.delete(`confirm:${code}`),
-        events.log({ type: 'email-change', oldEmail, newEmail: confirmObj.email }),
+        events.log({
+            type: 'email-change',
+            oldEmail,
+            newEmail: confirmObj.email,
+        }),
     ]);
 };
 
@@ -208,5 +229,8 @@ UserEmail.confirmByUid = async function (uid) {
         user.email.expireValidation(uid),
         user.reset.cleanByUid(uid),
     ]);
-    await plugins.hooks.fire('action:user.email.confirmed', { uid: uid, email: currentEmail });
+    await plugins.hooks.fire('action:user.email.confirmed', {
+        uid: uid,
+        email: currentEmail,
+    });
 };
